added the working check for support and some minor other things
Added the working check for support since there are browsers who
implemented it natively.
<style scoped> and <style scoped="scoped"> are both considered scoped
like it should since they are semantically equivalent.
Added a log statement to indicate that the current browser does not
support scoped when no support is found.
Removed the unnecessary toLowerCase() which was executed on the
node.nodeName.
And replaced the new RexExp() with the shorthand notiation